pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-search"
version = "0.1.0"
description = "Topic queries → web search → scrape → LLM-structured tables with provenance"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "trafilatura>=1.6",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.6"]

[project.scripts]
agentic-search = "agentic_search.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agentic_search = ["static/*.html"]

[tool.ruff]
line-length = 100
target-version = "py310"
